requirements.txt
absl-py
dm-haiku
jax
jaxtyping
matplotlib
ml_collections
optax
pandas
pydantic
shortuuid
tqdm
wandb
